requirements.txt
WTForms==2.1
pronouncing==0.1.5
Flask_Bootstrap==3.3.7.1
Flask==0.12.2
certifi==2017.4.17
Cython==0.25.2
pyOpenSSL==17.1.0
backports.functools_lru_cache==1.4
ed25519ll==0.6
protobuf==3.3.0
ipaddr==2.1.11
jsonschema==2.6.0
keyring==10.4.0
lxml==3.8.0
ndg_httpsclient==0.4.2
numpy==1.13.1
pyasn1==0.2.3
pytest==3.1.3
redis==2.10.5
wincertstore==0.2
pyxdg==0.25
PyYAML==3.12
